minor re-arrangement
